Добавил обработку исключений при некорректном вводе, при делении на ноль, при передачи нулевого указателя в качестве параметра

+++ Circle.h
@@ -3,12 +3,12 @@
 //класс окружность
 class Circle {
     Point Center; //центр окружности
-    int R; //радиус
+    double R; //радиус
 public:
     Circle(); //конструктор без параметров
-    Circle(int x, int y, int R); //конструктор с параметрами
-    Circle(int x, int y, int R, string metka); //конструктор с параметрами
-    void Init(int x, int y, int R); //метод инициализации
+    Circle(int x, int y, double R); //конструктор с параметрами
+    Circle(int x, int y, double R, string metka); //конструктор с параметрами
+    void Init(int x, int y, double R); //метод инициализации
     void Display(); //метод вывода
     void Read(); //метод ввода
     void PrintEquation(); //метод вывода уравнения окружности
